Module React

The React library

type domRef

type 'value ref = {

mutable current : 'value;

}

module Ref : sig ... end

val createRef : unit -> 'a option ref

val useRef : 'a -> 'a ref

val forwardRef : (unit -> 'a) -> 'a

module Event : sig ... end

module JSX : sig ... end

All of those types are used by the server-reason-react.ppx internally to represent valid React code from the server. It currently different from reason-react-ppx due to a need for knowing the types since ReactDOM needs to render differently depending on the type.

type element =

| Lower_case_element of {

key : string option;

tag : string;

attributes : JSX.prop list;

children : element list;

}

| Upper_case_component of unit -> element

| Async_component of unit -> element Lwt.t

| Client_component of {

props : client_props;

client : element;

import_module : string;

import_name : string;

}

| List of element array

| Text of string

| InnerHtml of string

| Fragment of element

| Empty

| Provider of element

| Consumer of element

| Suspense of {

key : string option;

children : element;

fallback : element;

}

and client_props = (string * client_prop) list

and client_prop =

| Json : Yojson.Basic.t -> client_prop

| Element : element -> client_prop

| Promise : 'a Js.Promise.t * ('a -> Yojson.Basic.t) -> client_prop

exception Invalid_children of string

module Fragment : sig ... end

val createElement : string -> JSX.prop list -> element list -> element

val createElementWithKey : ?key:string option -> string -> JSX.prop list -> element list -> element

val fragment : element -> element

val cloneElement : element -> JSX.prop list -> element

val string : string -> element

val null : element

val int : int -> element

val float : float -> element

val array : element array -> element

val list : element list -> element

type 'a provider = value:'a -> children:element -> unit -> element

type 'a context = {

current_value : 'a ref;

provider : 'a provider;

consumer : children:element -> element;

}

module Context : sig ... end

val createContext : 'a -> 'a Context.t

module Suspense : sig ... end

type any_promise =

| Any_promise : 'a Lwt.t -> any_promise

exception Suspend of any_promise

val useContext : 'a Context.t -> 'a

val useState : (unit -> 'state) -> 'state * (('state -> 'state) -> unit)

val useMemo : (unit -> 'a) -> 'a

val useMemo0 : (unit -> 'a) -> 'a

val useMemo1 : (unit -> 'a) -> 'b -> 'a

val useMemo2 : (unit -> 'a) -> 'b -> 'a

val useMemo3 : (unit -> 'a) -> 'b -> 'a

val useMemo4 : (unit -> 'a) -> 'b -> 'a

val useMemo5 : (unit -> 'a) -> 'b -> 'a

val useMemo6 : (unit -> 'a) -> 'b -> 'a

val useCallback : 'a -> 'a

val useCallback0 : 'a -> 'a

val useCallback1 : 'a -> 'b -> 'a

val useCallback2 : 'a -> 'b -> 'a

val useCallback3 : 'a -> 'b -> 'a

val useCallback4 : 'a -> 'b -> 'a

val useCallback5 : 'a -> 'b -> 'a

val useCallback6 : 'a -> 'b -> 'a

val useId : unit -> string

type ('input, 'output) callback = 'input -> 'output

val useSyncExternalStore : subscribe:((unit -> unit) -> (unit, unit) callback) -> getSnapshot:(unit -> 'snapshot) -> 'snapshot

deprecated Use useSyncExternalStoreWithServer instead. More info at https://github.com/facebook/react/blob/603e6108f39c6663ec703eed34a89ff1bf0cb70c/packages/react-server/src/ReactFizzHooks.js#L561-L566

val useSyncExternalStoreWithServer : subscribe:((unit -> unit) -> (unit, unit) callback) -> getSnapshot:(unit -> 'snapshot) -> getServerSnapshot:(unit -> 'snapshot) -> 'snapshot

val useReducer : ('state -> 'action -> 'state) -> 'state -> 'state * ('action -> unit)

val useReducerWithMapState : ('state -> 'action -> 'initialState) -> 'initialState -> ('initialState -> 'state) -> 'state * ('action -> unit)

val useEffect : (unit -> (unit -> unit) option) -> unit

val useEffect0 : (unit -> (unit -> unit) option) -> unit

val useEffect1 : (unit -> (unit -> unit) option) -> 'dependency array -> unit

val useEffect2 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2) -> unit

val useEffect3 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3) -> unit

val useEffect4 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3 * 'dependency4) -> unit

val useEffect5 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3 * 'dependency4 * 'dependency5) -> unit

val useEffect6 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3 * 'dependency4 * 'dependency5 * 'dependency6) -> unit

val useLayoutEffect0 : (unit -> (unit -> unit) option) -> unit

val useLayoutEffect1 : (unit -> (unit -> unit) option) -> 'dependency array -> unit

val useLayoutEffect2 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2) -> unit

val useLayoutEffect3 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3) -> unit

val useLayoutEffect4 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3 * 'dependency4) -> unit

val useLayoutEffect5 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3 * 'dependency4 * 'dependency5) -> unit

val useLayoutEffect6 : (unit -> (unit -> unit) option) -> ('dependency1 * 'dependency2 * 'dependency3 * 'dependency4 * 'dependency5 * 'dependency6) -> unit

val setDisplayName : 'component -> string -> unit

module Children : sig ... end

val suspend : 'a Lwt.t -> unit

module Experimental : sig ... end

val useTransition : unit -> bool * ((unit -> unit) -> unit)

val useDebugValue : 'value -> ?format:('value -> string) -> unit

val useDeferredValue : 'value -> 'value