Module React.Step

Update steps.

Update functions returned by S.create and E.create implicitely create and execute update steps when used without specifying their step argument.

Using explicit step values with these functions gives more control on the time when the update step is perfomed and allows to perform simultaneous primitive signal updates and event occurences. See also the documentation about update steps and simultaneous events.

Steps

type t = step

The type for update steps.

val create : unit -> step

create () is a new update step.

val execute : step -> unit

execute step executes the update step.

raises Invalid_argument
if step was already executed.